Make setsockopt properly check that values are greater than 0
The XSLT transform had a minor typo that I noticed while I was re-writing the tranform to port it to HHVM, this typo causes the check to ensure that the input value is greater than or equal to zero to never succeed, causing the check to never be generated.
This simply fixes the typo. It does not regenerate the generated file.