sockopts: code condensed by moving to makros
The code is autogenerated by the *.gsl scripts. In order to
minimize the c code now parts of the gsl code generation is
moved to new makro definitions in zmq_sockopts_makros.h

This reduces the overall lines of the file and makes it more
readable for future additions of flags.